+++ app/assets/javascripts/ci/merge_requests/components/pipelines_table_wrapper.vue
@@ -337,7 +337,7 @@ export default {
     },
   ),
   userPermissionsDocsPath: helpPagePath('user/permissions.md', {
-    anchor: 'cicd',
+    anchor: 'project-cicd',
   }),
 };
 </script>