+++ app/assets/javascripts/entrypoints/lookbook/experiment_badge.js
@@ -1,0 +1,3 @@
+import { initPopovers } from '~/popovers';
+
+initPopovers();