+++ app/assets/javascripts/jira_connect/branches/components/new_branch_form.vue
@@ -84,7 +84,7 @@ export default {
       this.displayAlert({
         message: I18N_NEW_BRANCH_PERMISSION_ALERT,
         variant: 'warning',
-        link: helpPagePath('user/permissions', { anchor: 'project-members-permissions' }),
+        link: helpPagePath('user/permissions', { anchor: 'project-permissions' }),
         dismissible: false,
       });
     },