+++ app/assets/javascripts/merge_requests/list/components/merge_requests_list_app.vue
@@ -82,7 +82,7 @@ import {
 } from '~/issues/list/constants';
 import CiIcon from '~/vue_shared/components/ci_icon/ci_icon.vue';
 import setSortPreferenceMutation from '~/issues/dashboard/queries/set_sort_preference.mutation.graphql';
-import issuableEventHub from '~/issues/list/eventhub';
+import issuableEventHub from '~/merge_requests/list/eventhub';
 import searchLabelsQuery from '~/issues/list/queries/search_labels.query.graphql';
 import { AutocompleteCache } from '../../utils/autocomplete_cache';
 import { i18n, BRANCH_LIST_REFRESH_INTERVAL } from '../constants';
@@ -790,7 +790,7 @@ export default {
     },
     async handleBulkUpdateClick() {
       if (!this.hasInitBulkEdit) {
-        const bulkUpdateSidebar = await import('~/issuable');
+        const bulkUpdateSidebar = await import('~/merge_requests/list');
         bulkUpdateSidebar.initBulkUpdateSidebar('issuable_');
 
         this.hasInitBulkEdit = true;