+++ app/components/rapid_diffs/viewers/text/parallel_hunk_component.rb
@@ -23,13 +23,11 @@ module RapidDiffs
               sides: [
                 {
                   line: pair[:left],
-                  position: :old,
-                  border: :right
+                  position: :old
                 },
                 {
                   line: pair[:right],
-                  position: :new,
-                  border: :both
+                  position: :new
                 }
               ],
               expanded: pair[:left]&.expanded? || pair[:right]&.expanded?