+++ app/helpers/time_helper.rb
@@ -17,18 +17,6 @@ module TimeHelper
     time_parts.map { |unit, value| TIME_UNIT_TRANSLATION[unit].call(value) }.to_sentence
   end
 
-  def duration_in_numbers(duration_in_seconds)
-    seconds = duration_in_seconds % 1.minute
-    minutes = (duration_in_seconds / 1.minute) % (1.hour / 1.minute)
-    hours = duration_in_seconds / 1.hour
-
-    if hours == 0
-      "%02d:%02d" % [minutes, seconds]
-    else
-      "%02d:%02d:%02d" % [hours, minutes, seconds]
-    end
-  end
-
   def time_in_milliseconds
     (Time.now.to_f * 1000).to_i
   end