+++ spec/frontend/ci/merge_requests/components/pipelines_table_wrapper_spec.js
@@ -198,7 +198,7 @@ describe('PipelinesTableWrapper component', () => {
           '/help/ci/pipelines/merge_request_pipelines.md#prerequisites',
         );
         expect(findUserPermissionsDocsLink().attributes('href')).toBe(
-          '/help/user/permissions.md#cicd',
+          '/help/user/permissions.md#project-cicd',
         );
 
         expect(findEmptyState().text()).toContain('To run a merge request pipeline');