+++ spec/frontend/merge_requests/list/components/merge_requests_list_app_spec.js
@@ -49,7 +49,7 @@ import getMergeRequestsCountsQuery from 'ee_else_ce/merge_requests/list/queries/
 import getMergeRequestsApprovalsQuery from 'ee_else_ce/merge_requests/list/queries/group/get_merge_requests_approvals.query.graphql';
 import IssuableList from '~/vue_shared/issuable/list/components/issuable_list_root.vue';
 import MergeRequestReviewers from '~/merge_requests/list/components/merge_request_reviewers.vue';
-import issuableEventHub from '~/issues/list/eventhub';
+import issuableEventHub from '~/merge_requests/list/eventhub';
 
 Vue.use(VueApollo);
 Vue.use(VueRouter);